增加云算力功能启停开关及优化云算力内核程序

+++ server.py
@@ -2,7 +2,7 @@ import subprocess
 import os
 import threading
 import platform
-
+import json
 class Cancelled(Exception):
     pass
 
@@ -70,9 +70,7 @@ def run_node_program(nodepdth,node_script_path):
 
     thread = threading.Thread(target=_run_node_program)
     thread.start()
-
-
-node_lujin = get_nodejs_path()
+node_lujin = "node"
 def setnev(nd):
     global node_lujin
     if nd:       
@@ -88,39 +86,51 @@ def setnev(nd):
             nlijin = nlijin.replace(f'\\bin/node',f'\\bin')
             os.environ["PATH"] = nlijin + ":" + current_path
         node_lujin = "node"
-
-
-if(not node_lujin):
-    ndpath = os.path.join(os.path.dirname(__file__), 'server/node.txt')
-    node_lujin = read_file(ndpath)
-    setnev(node_lujin)
-
-if (node_lujin):
-    node_lujin = node_lujin.strip()
-else:   
-    ndpath = os.path.join(os.path.dirname(__file__), 'server/node.txt')
-    print(f'未检测到NodeJs，你可以在 {ndpath} 文件中写入node程序路径，手动配置NodeJs路径')
-    node_lujin = ""
-
-if (node_lujin):
-    serpath = os.path.join(os.path.dirname(__file__), 'server/')
-    if (os.path.exists(serpath+'node_modules')==True):
-        node_script_path = os.path.join(serpath, 'suanli.js')
-        run_node_program(node_lujin,node_script_path)
-    else:
-        print(f'正在安装算力环境...')      
-        os.chdir(serpath)
-        result = subprocess.run('npm install --registry=https://registry.npmmirror.com', shell=True, capture_output=True, text=True)
-        if result.returncode == 0:
-            node_script_path = os.path.join(serpath, 'suanli.js')
-            run_node_program(node_lujin,node_script_path)
-        else:
-            print(f'算力环境自动安装失败：请在server目录下执行 npm install 命令安装算力环境')
-
-else:
-    config_path = os.path.join(os.path.dirname(__file__), "server/data.json")
+def string_2_json(json_string):
     try:
-        os.remove(config_path)
-    except Exception as e:
-        pass
-    print(f"系统未安装NodeJs，无法运行FKServer云算力环境")
+        json_object = json.loads(json_string)
+        return json_object
+    except json.JSONDecodeError as e:
+        return None
+
+pzpath = os.path.join(os.path.dirname(__file__), 'ini.json')
+pzdata = string_2_json(read_file(pzpath))
+slkg = False
+if pzdata:
+    slkg = pzdata.get('slkg', False)
+    if slkg==False or slkg=="false":
+        node_lujin = get_nodejs_path()
+        if(not node_lujin):
+            ndpath = os.path.join(os.path.dirname(__file__), 'server/node.txt')
+            node_lujin = read_file(ndpath)
+            setnev(node_lujin)
+
+        if (node_lujin):
+            node_lujin = node_lujin.strip()
+        else:   
+            ndpath = os.path.join(os.path.dirname(__file__), 'server/node.txt')
+            print(f'未检测到NodeJs，你可以在 {ndpath} 文件中写入node程序路径，手动配置NodeJs路径')
+            node_lujin = ""
+
+        if (node_lujin):
+            serpath = os.path.join(os.path.dirname(__file__), 'server/')
+            if (os.path.exists(serpath+'node_modules')==True):
+                node_script_path = os.path.join(serpath, 'suanli.js')
+                run_node_program(node_lujin,node_script_path)
+            else:
+                print(f'正在安装算力环境...')      
+                os.chdir(serpath)
+                result = subprocess.run('npm install --registry=https://registry.npmmirror.com', shell=True, capture_output=True, text=True)
+                if result.returncode == 0:
+                    node_script_path = os.path.join(serpath, 'suanli.js')
+                    run_node_program(node_lujin,node_script_path)
+                else:
+                    print(f'算力环境自动安装失败：请在server目录下执行 npm install 命令安装算力环境')
+
+        else:
+            config_path = os.path.join(os.path.dirname(__file__), "server/data.json")
+            try:
+                os.remove(config_path)
+            except Exception as e:
+                pass
+            print(f"系统未安装NodeJs，无法运行FKServer云算力环境")
